refactor(web): polish dir structure
